+++ src/draw_world.cpp
@@ -721,7 +721,7 @@ void draw_coll_surfaces(bool draw_trans, bool reflection_pass) {
 				}
 				c.draw_cobj(cix, last_tid, last_group_id, s, cdb);
 				if (using_lt_atten) {cdb.flush();} // must flush because ulocs[2] is per-cube
-				assert(cix == ix); // should not have changed
+				assert((int)cix == ix); // should not have changed
 			}
 		} // for i
 		if (in_portal) {portal::post_draw(portal_verts);}
